+++ packages/jssdk/src/frame/slider.ts
@@ -171,7 +171,7 @@ export class SliderManager {
    * @todo test this and remove
    */
   // async showAppForm(params: any): Promise<void> {
-  //   console.warn(`@deprecated showAppForm`)
+  //   console.warn(`deprecated showAppForm`)
   //   return this.#messageManager.send(MessageCommands.showAppForm, {
   //     params: params,
   //     isSafely: true