+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * feat(PlacementManager): use the property name `placement` instead of `title`
 * feat(apiVersion): support api 3
 * feat(Logger): add new logger system
+* fix(logger\LoggerBrowser): marked as deprecated
 * fix(tools\Type|Text|Browser): improve export
 * feat(tools\environment): added tool for environment detection
 * feat(logger\handler\TelegramHandler): add Telegram handler
@@ -22,6 +23,8 @@
 * feat(core): add tools and actions
 * fix(RefreshTokenError): extends SdkError
 * fix(frame\SliderManager): remove showAppForm
+* fix(frame\DialogManager): remove deprecated marker
+
 
 ## [2.1.0](https://github.com/bitrix24/b24jssdk/compare/v0.5.1...v1.1.0) (2025-12-xx)
 ## [1.1.0](https://github.com/bitrix24/b24jssdk/compare/v0.5.1...v1.1.0) (2025-12-xx)